+++ src/commands/metadata/mosaic.ts
@@ -20,14 +20,12 @@ import {command, metadata, option} from 'clime';
 import {Metadata, MetadataHttp} from 'nem2-sdk';
 import {ProfileCommand, ProfileOptions} from '../../profile.command';
 import {MosaicIdResolver} from '../../resolvers/mosaic.resolver';
-import {MosaicIdValidator} from '../../validators/mosaicId.validator';
 import {MetadataEntryTable} from './account';
 
 export class CommandOptions extends ProfileOptions {
     @option({
         flag: 'm',
         description: 'Mosaic id in hexadecimal format.',
-        validator: new MosaicIdValidator(),
     })
     mosaicId: string;
 }
@@ -45,9 +43,9 @@ export default class extends ProfileCommand {
     execute(options: CommandOptions) {
         this.spinner.start();
         const profile = this.getProfile(options);
-        const metadataHttp = new MetadataHttp(profile.url);
         const mosaicId = new MosaicIdResolver().resolve(options);
 
+        const metadataHttp = new MetadataHttp(profile.url);
         metadataHttp.getMosaicMetadata(mosaicId)
             .subscribe((metadataEntries) => {
                 this.spinner.stop(true);
@@ -61,10 +59,8 @@ export default class extends ProfileCommand {
                 }
             }, (err) => {
                 this.spinner.stop(true);
-                let text = '';
-                text += chalk.red('Error');
                 err = err.message ? JSON.parse(err.message) : err;
-                console.log(text, err.body && err.body.message ? err.body.message : err);
+                console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
             });
     }
 }