+++ src/commands/monitor/unconfirmed.ts
@@ -33,25 +33,20 @@ export default class extends MonitorAddressCommand {
     @metadata
     execute(options: MonitorAddressOptions) {
         const profile = this.getProfile(options);
-        const listener = new Listener(profile.url);
         const address = new AddressResolver().resolve(options, profile);
 
         console.log(chalk.green('Monitoring ') + `${address.pretty()} using ${profile.url}`);
-
+        const listener = new Listener(profile.url);
         listener.open().then(() => {
             listener.unconfirmedAdded(address).subscribe((transaction) => {
                 console.log('\n' + this.transactionService.formatTransactionToFilter(transaction));
             }, (err) => {
-                let text = '';
-                text += chalk.red('Error');
                 err = err.message ? JSON.parse(err.message) : err;
-                console.log(text, err.body && err.body.message ? err.body.message : err);
+                console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
             });
         }, (err) => {
-            let text = '';
-            text += chalk.red('Error');
             err = err.message ? JSON.parse(err.message) : err;
-            console.log(text, err.body && err.body.message ? err.body.message : err);
+            console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
         });
     }
 }